+++ Altinn/AT.Common.Altinn.Publish/Extensions/InstanceExtensions.cs
@@ -10,28 +10,25 @@ public static class InstanceExtensions
     /// <summary>
     /// Extracts the AltinnInstance Guid from the AltinnInstance Id
     /// </summary>
-    /// <param name="AltinnInstance"></param>
+    /// <param name="altinnInstance">The instance whose identifier to parse.</param>
     /// <returns></returns>
     /// <exception cref="InvalidOperationException"></exception>
-    public static Guid GetInstanceGuid(this AltinnInstance AltinnInstance)
+    public static Guid GetInstanceGuid(this AltinnInstance altinnInstance)
     {
-        // Split the Id by '/' and parse the second part as a Guid
-        if (AltinnInstance.Id.Split("/").Length != 2)
+        if (altinnInstance.Id?.Split('/') is not [_, var instanceId])
         {
             throw new InvalidOperationException(
                 "AltinnInstance ID must be in the format partyId/instanceGuid"
             );
         }
 
-        // Ensure the second part is a valid Guid
-        if (!Guid.TryParse(AltinnInstance.Id.Split("/")[1], out var instanceGuid))
+        if (!Guid.TryParse(instanceId, out var instanceGuid))
         {
             throw new InvalidOperationException(
                 "AltinnInstance ID must contain a valid Guid in the second part"
             );
         }
 
-        // Return the parsed Guid
         return instanceGuid;
     }
 }