+++ Altinn/AT.Common.Altinn.Publish/Model/Api/Request/CorrespondenceDtos.cs
@@ -21,18 +21,6 @@ public record BaseCorrespondence
     public required string ResourceId { get; init; }
 
     /// <summary>
-    /// The Sending organization of the correspondence.
-    /// </summary>
-    /// <remarks>
-    /// Organization number must be formatted as countrycode:organizationnumber.
-    /// </remarks>
-    [JsonPropertyName("sender")]
-    [Obsolete(
-        "Sender is deprecated and will be removed in a future version. The sender is now automatically determined from the Resource Registry based on the resourceId."
-    )]
-    public string? Sender { get; init; }
-
-    /// <summary>
     /// A reference used by senders and receivers to identify a specific Correspondence using external identification methods.
     /// </summary>
     [JsonPropertyName("sendersReference")]